First attempt to write an API for autocomplete
This change adds the `suggestive.Suggestive` class, that allows us to declare
new autocomplete namespaces. Instances of this class represents the `db` they
take as the first parameter. This way, you can have separate data sets for
different features in your projects.

This is a more or less `document` oriented indexing system. The documents are
represented with python dictionaries and all the fields are searchable. It may
change when we add the option for indexing data to optimize the search.