Cleaning up the document before indexing it again
We would create garbage in the term set in a specific corner case. Consider
the following code:

  >>> backend.index([{"id": 0, "name": "lincoln"}], score="id")

When you index anything with the same ID, like this:

  >>> backend.index([{"id": 0, "name": "mingwei"}], score="id")

Then you will have garbage in your term list. If you look for the
string "Lin", you will find "Mingwei" which does not make sense