Removing the `db` param from the `Suggestive` constructor
I was thinking about adding a switcher for multiple dbs, but I think
that it is actually easier to do this by introducing a new keyword
argument that receives a uri. So the user would have to pass either a
backend or a uri, that would be resolved to a backend and the data
in this uri would be used as params to the backend constructor.